[BPK-2693] [BPK-2693] [BPK-2693] Split out docekd inputs
